Registry: refactor deployment/configmap allowing storage override
We had to refactor the way we were configuring the container to deal with changing the `storage` item of the Registry configuration.

The container ships with a `/etc/docker/registry/config.yml` that provides:
```
storage:
  filesystem:
...
```

There is no way to _remove_ this by adding/removing environment variables, so we instead needed to override it. To do that, we have refactored the ConfigMap to have a single value that is then `volumeMount`'d  into the container, thus overriding.

I have provided for the matching defaults from the `Dockerfile` of the `registry:2.6` at the Docker Hub.